Room Painting
This program reads two types of input:
The number of available paint can sizes and the number of paint needs.

It then reads in the actual available cans and the needed cans. For each paint need, the program finds the smallest available can that can fulfill the need (available can  ≥ need). It calculates and adds the difference (waste) between the available can and the actual need. If no can is big enough for a particular need, it skips that need. Finally, the program outputs the total amount of wasted paint.